refactor: move seriesdetail to the app-series module
The first series part to move across is the SeriesDetail
