+++ app-series/build.gradle
@@ -37,6 +37,7 @@ dependencies {
     implementation project(path: ':account')
     implementation project(path: ':core')
     implementation project(path: ':series')
+    implementation project(path: ':server')
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation "androidx.appcompat:appcompat:$appcompat_version"
@@ -46,10 +47,12 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
     implementation "com.chesire:lifecyklelog:$lifecykle_version"
     implementation "com.github.bumptech.glide:glide:$glide_version"
+    implementation "com.github.hadilq.liveevent:liveevent:$liveevent_version"
     implementation "com.google.dagger:dagger:$dagger_version"
     implementation "com.google.dagger:dagger-android:$dagger_version"
     implementation "com.google.dagger:dagger-android-support:$dagger_version"
     implementation "com.google.android.material:material:$material_version"
+    implementation "com.jakewharton.timber:timber:$timber_version"
 
     testImplementation project(path: ':testing')
     testImplementation 'androidx.arch.core:core-testing:2.1.0-rc01'
