+++ app-series/src/main/java/com/chesire/malime/app/series/detail/SeriesDetailError.kt
@@ -1,4 +1,4 @@
-package com.chesire.malime.flow.series.detail
+package com.chesire.malime.app.series.detail
 
 /**
  * List of errors for use with the SeriesDetail flow.
