+++ app-series/src/main/res/layout/fragment_series_detail.xml
