+++ app-series/src/main/res/layout/view_series_detail_series_status.xml
