+++ app/build.gradle
@@ -88,7 +88,7 @@ dependencies {
     implementation "com.afollestad.material-dialogs:lifecycle:$material_dialog_version"
     implementation "com.chesire:lifecyklelog:$lifecykle_version"
     implementation "com.github.bumptech.glide:glide:$glide_version"
-    implementation 'com.github.hadilq.liveevent:liveevent:1.0.1'
+    implementation "com.github.hadilq.liveevent:liveevent:$liveevent_version"
     implementation "com.google.android.material:material:$material_version"
     implementation "com.google.dagger:dagger:$dagger_version"
     implementation "com.google.dagger:dagger-android:$dagger_version"
