+++ app/src/test/java/com/chesire/malime/flow/series/list/SeriesListViewModelTests.kt
@@ -2,7 +2,7 @@ package com.chesire.malime.flow.series.list
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
-import com.chesire.malime.AuthCaster
+import com.chesire.malime.core.AuthCaster
 import com.chesire.malime.testing.CoroutinesMainDispatcherRule
 import com.chesire.malime.core.flags.AsyncState
 import com.chesire.malime.core.flags.UserSeriesStatus
