+++ build.gradle
@@ -71,6 +71,7 @@ ext {
     glide_version = "4.10.0"
     lifecycle_version = "2.2.0-beta01"
     lifecykle_version = "2.0.0"
+    liveevent_version = "1.0.1"
     material_version = "1.1.0-beta01"
     moshi_version = "1.8.0"
     preferences_version = "1.1.0"
