+++ core/src/main/java/com/chesire/malime/core/AuthCaster.kt
@@ -1,4 +1,4 @@
-package com.chesire.malime
+package com.chesire.malime.core
 
 import javax.inject.Inject
 import javax.inject.Singleton
